@@ -11,16 +11,17 @@
         }
 	],
     "require": {
-        "php": ">=5.5.0",
-        "doctrine/instantiator": "~1.0.5",
-        "ramsey/uuid": "~3.5.1"
+        "php": ">=7.1",
+        "doctrine/instantiator": "^1.0",
+        "ramsey/uuid": "^3.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^4.8 || ^5.6",
-        "php-amqplib/php-amqplib": "~2.6.3",
-        "predis/predis": "~1.1.1",
-        "drealecs/time-benchmark": "~1.0.0",
-        "symfony/process": "~3.1.5"
+        "phpunit/phpunit": "^7.0",
+        "php-amqplib/php-amqplib": "^2.6",
+        "predis/predis": "^1.1",
+        "ext-redis": "*",
+        "drealecs/time-benchmark": "~1.0",
+        "symfony/process": "^4.2"
     },
     "suggest": {
       "php-amqplib/php-amqplib": "Provides rabbitmq PHP implementation",
