@@ -1,7 +1,10 @@
 <?php
+declare(strict_types=1);
+
 namespace ParallelTask;
 
-use ParallelTask\Queue\Queue;
+use ParallelTask\Queue\ConsumeQueue;
+use ParallelTask\Queue\PublishQueue;
 use ParallelTask\Task\TaskFactory;
 use ParallelTask\Task\TaskFactorySimple;
 use ParallelTask\Task\TaskInputMessageTransformer;
@@ -14,8 +17,10 @@ use ParallelTask\Task\TaskScheduler;
 
 class ExecutorWorkerBuilder
 {
-    /** @var Queue */
-    private $queue;
+    /** @var PublishQueue */
+    private $publishQueue;
+    /** @var ConsumeQueue */
+    private $consumeQueue;
     /** @var TaskInputMessageTransformer */
     private $taskInputMessageTransformer;
     /** @var TaskResultMessageTransformer */
@@ -32,84 +37,61 @@ class ExecutorWorkerBuilder
         $this->taskRunnerSupervisor = new TaskRunnerNullSupervisor();
     }
 
-    /**
-     * @param Queue $queue
-     * @return ExecutorWorkerBuilder
-     */
-    public function withQueue(Queue $queue)
+    public function withPublishQueue(PublishQueue $queue): ExecutorWorkerBuilder
     {
-        $this->queue = $queue;
+        $this->publishQueue = $queue;
         return $this;
     }
 
-    /**
-     * @param TaskInputMessageTransformer $taskMessageTransformer
-     * @return ExecutorWorkerBuilder
-     */
-    public function withTaskInputMessageTransformer(TaskInputMessageTransformer $taskMessageTransformer)
+    public function withConsumeQueue(ConsumeQueue $queue): ExecutorWorkerBuilder
+    {
+        $this->consumeQueue = $queue;
+        return $this;
+    }
+
+    public function withTaskInputMessageTransformer(TaskInputMessageTransformer $taskMessageTransformer): ExecutorWorkerBuilder
     {
         $this->taskInputMessageTransformer = $taskMessageTransformer;
         return $this;
     }
 
-    /**
-     * @param TaskResultMessageTransformer $taskMessageTransformer
-     * @return ExecutorWorkerBuilder
-     */
-    public function withTaskResultMessageTransformer(TaskResultMessageTransformer $taskMessageTransformer)
+    public function withTaskResultMessageTransformer(TaskResultMessageTransformer $taskMessageTransformer): ExecutorWorkerBuilder
     {
         $this->taskResultMessageTransformer = $taskMessageTransformer;
         return $this;
     }
 
-    /**
-     * @param TaskFactory $taskFactory
-     * @return ExecutorWorkerBuilder
-     */
-    public function withTaskFactory(TaskFactory $taskFactory)
+    public function withTaskFactory(TaskFactory $taskFactory): ExecutorWorkerBuilder
     {
         $this->taskFactory = $taskFactory;
         return $this;
     }
 
-    /**
-     * @param TaskRunnerSupervisor $taskRunnerSupervisor
-     * @return ExecutorWorkerBuilder
-     */
-    public function withTaskRunnerSupervisor(TaskRunnerSupervisor $taskRunnerSupervisor)
+    public function withTaskRunnerSupervisor(TaskRunnerSupervisor $taskRunnerSupervisor): ExecutorWorkerBuilder
     {
         $this->taskRunnerSupervisor = $taskRunnerSupervisor;
         return $this;
     }
 
-    /**
-     * @return Executor
-     */
-    public function buildExecutor()
+    public function buildExecutor(): Executor
     {
-        $this->checkRequiredParameters();
+        if (!$this->publishQueue instanceof PublishQueue) {
+            throw new \RuntimeException('No queue was set. Use withPublishQueue($queue) method to set one before building');
+        }
 
-        $taskScheduler = new TaskScheduler($this->queue, $this->taskInputMessageTransformer, $this->taskResultMessageTransformer);
+        $taskScheduler = new TaskScheduler($this->publishQueue, $this->taskInputMessageTransformer, $this->taskResultMessageTransformer);
 
         return new Executor($taskScheduler);
     }
 
-    /**
-     * @return Worker
-     */
-    public function buildWorker()
+    public function buildWorker(): Worker
     {
-        $this->checkRequiredParameters();
+        if (!$this->consumeQueue instanceof ConsumeQueue) {
+            throw new \RuntimeException('No queue was set. Use withConsumeQueue($queue) method to set one before building');
+        }
 
-        $taskRunner = new TaskRunner($this->queue, $this->taskInputMessageTransformer, $this->taskFactory, $this->taskResultMessageTransformer, $this->taskRunnerSupervisor);
+        $taskRunner = new TaskRunner($this->consumeQueue, $this->taskInputMessageTransformer, $this->taskFactory, $this->taskResultMessageTransformer, $this->taskRunnerSupervisor);
 
         return new Worker($taskRunner);
-    }
-
-    private function checkRequiredParameters()
-    {
-        if (!$this->queue instanceof Queue) {
-            throw new \Exception('No queue was set. Use withQueue($queue) method to set one before building');
-        }
     }
 }
