@@ -1,19 +1,23 @@
 <?php
+declare(strict_types=1);
+
 namespace ParallelTask\Queue\RabbitMQ;
 
+use ParallelTask\Queue\ConsumeQueue;
 use ParallelTask\Queue\InputMessage;
 use ParallelTask\Queue\InputMessageIdentifier;
 use ParallelTask\Queue\OutputMessage;
-use ParallelTask\Queue\Queue;
+use ParallelTask\Queue\PublishQueue;
+use PhpAmqpLib\Channel\AMQPChannel;
 use PhpAmqpLib\Connection\AMQPStreamConnection;
 use PhpAmqpLib\Message\AMQPMessage;
 use Ramsey\Uuid\Uuid;
 
-class RabbitMQQueue implements Queue
+class RabbitMQQueue implements PublishQueue, ConsumeQueue
 {
     /** @var AMQPStreamConnection */
     private $connection;
-    /** @var \AMQPChannel */
+    /** @var AMQPChannel */
     private $channel;
     /** @var string */
     private $replyQueueName;
@@ -30,7 +34,7 @@ class RabbitMQQueue implements Queue
         $this->channel = $connection->channel();
     }
 
-    public function putInput($type, InputMessage $inputMessage)
+    public function putInput(string $type, InputMessage $inputMessage): void
     {
         $this->declareQueue($type);
 
@@ -38,7 +42,7 @@ class RabbitMQQueue implements Queue
         $this->channel->basic_publish($amqpMessage, $this->getExchangeName($type));
     }
 
-    public function submitInput($type, InputMessage $inputMessage)
+    public function submitInput(string $type, InputMessage $inputMessage): InputMessageIdentifier
     {
         $this->declareQueue($type);
 
@@ -64,7 +68,7 @@ class RabbitMQQueue implements Queue
         return new InputMessageIdentifier($id);
     }
 
-    public function run($type, callable $runCallback)
+    public function run(string $type, callable $runCallback): void
     {
         static $wrapperPassingRunCallback;
         if (!isset($wrapperPassingRunCallback)) {
@@ -99,7 +103,7 @@ class RabbitMQQueue implements Queue
         unset($wrapperPassingRunCallback['runCallback']);
     }
 
-    public function getOutput($type, InputMessageIdentifier $identifier)
+    public function getOutput(string $type, InputMessageIdentifier $identifier): OutputMessage
     {
         $id = $identifier->getId();
 
