@@ -1,16 +1,19 @@
 <?php
+declare(strict_types=1);
+
 namespace ParallelTask\Queue\Redis;
 
 use ParallelTask\Queue\InputMessage;
 use ParallelTask\Queue\InputMessageIdentifier;
 use ParallelTask\Queue\InputMessageWithIdentifier;
-use ParallelTask\Queue\NonCallbackQueue;
+use ParallelTask\Queue\NonCallbackConsumeQueue;
 use ParallelTask\Queue\OutputMessage;
+use ParallelTask\Queue\PublishQueue;
 use Predis\Client;
 use Predis\ClientContextInterface;
 use Ramsey\Uuid\Uuid;
 
-class PredisQueue extends NonCallbackQueue
+class PredisQueue extends NonCallbackConsumeQueue implements PublishQueue
 {
     /**
      * @var int
@@ -27,11 +30,7 @@ class PredisQueue extends NonCallbackQueue
         $this->redis = $redis;
     }
 
-    /**
-     * @param $type
-     * @param InputMessage $inputMessage
-     */
-    public function putInput($type, InputMessage $inputMessage)
+    public function putInput(string $type, InputMessage $inputMessage): void
     {
         $id = Uuid::uuid4()->toString();
         $data = $inputMessage->getData();
@@ -39,12 +38,7 @@ class PredisQueue extends NonCallbackQueue
         $this->storeInput($type, $id, $data, false);
     }
 
-    /**
-     * @param $type
-     * @param InputMessage $inputMessage
-     * @return InputMessageIdentifier
-     */
-    public function submitInput($type, InputMessage $inputMessage)
+    public function submitInput(string $type, InputMessage $inputMessage): InputMessageIdentifier
     {
         $id = Uuid::uuid4()->toString();
         $data = $inputMessage->getData();
@@ -66,11 +60,7 @@ class PredisQueue extends NonCallbackQueue
         });
     }
 
-    /**
-     * @param string $type
-     * @return InputMessageWithIdentifier
-     */
-    public function getInput($type)
+    public function getInput(string $type): InputMessageWithIdentifier
     {
         if (rand(0, 99) == 0) {
             $this->requeueOldWorkingMessages($type);
@@ -91,12 +81,7 @@ class PredisQueue extends NonCallbackQueue
         return new InputMessageWithIdentifier($identifier, $inputMessage);
     }
 
-    /**
-     * @param string $type
-     * @param InputMessageIdentifier $identifier
-     * @param OutputMessage $outputMessage
-     */
-    public function putOutput($type, InputMessageIdentifier $identifier, OutputMessage $outputMessage)
+    public function putOutput(string $type, InputMessageIdentifier $identifier, OutputMessage $outputMessage): void
     {
         $messageId = $identifier->getId();
 
@@ -116,12 +101,7 @@ class PredisQueue extends NonCallbackQueue
         });
     }
 
-    /**
-     * @param string $type
-     * @param InputMessageIdentifier $identifier
-     * @return OutputMessage
-     */
-    public function getOutput($type, InputMessageIdentifier $identifier)
+    public function getOutput(string $type, InputMessageIdentifier $identifier): OutputMessage
     {
         $messageId = $identifier->getId();
 
