@@ -1,24 +1,29 @@
 <?php
+declare(strict_types=1);
+
 namespace ParallelTask\Task;
 
 final class FutureTaskResult
 {
+    /** @var TaskResult */
+    private $taskResult;
     /** @var \Closure */
     private $taskResultCallback;
+    /** @var bool */
+    private $taskResultCallbackSolved = false;
 
     public function __construct(\Closure $taskResultCallback)
     {
         $this->taskResultCallback = $taskResultCallback;
     }
 
-    /**
-     * @return TaskResult
-     */
-    public function getTaskResult()
+    public function getTaskResult(): TaskResult
     {
-        $taskResultCallback = $this->taskResultCallback;
-        $taskResult = $taskResultCallback();
+        if (!$this->taskResultCallbackSolved) {
+            $this->taskResult = ($this->taskResultCallback)();
+            $this->taskResultCallbackSolved = true;
+        }
 
-        return $taskResult;
+        return $this->taskResult;
     }
 }
