@@ -1,13 +1,15 @@
 <?php
+declare(strict_types=1);
+
 namespace ParallelTask\Task;
 
 final class TaskRunnerNullSupervisor implements TaskRunnerSupervisor
 {
-    public function markRunnerStart()
+    public function markRunnerStart(): void
     {
     }
 
-    public function shouldRunnerStop()
+    public function shouldRunnerStop(): bool
     {
         return false;
     }
